Add test for Mission and MissionCreate
